+++ services/diagnosis-agent/my_agent/agent.py
@@ -9,7 +9,7 @@ class ActionParams(BaseModel):
     target_revision: str = ""
     env_key: str = ""
     env_value: str = ""
-    memory_mb: float = 0
+    memory_mb: int = 0
 
 
 class RecoveryDecision(BaseModel):
